scheduler/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx>=0.27.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
redis>=5.0.0
# Observability
prometheus-client>=0.20.0
structlog>=24.1.0
opentelemetry-sdk>=1.25.0
opentelemetry-instrumentation-fastapi>=0.46b0
opentelemetry-instrumentation-httpx>=0.46b0
opentelemetry-exporter-otlp-proto-grpc>=1.25.0
